Add kinesis exporter (#2)
This adds our existing OC Kinesis exporter as an OT exporter

For now this directly maps the existing OC kinesis exporter to a new OT
one. I've tried to not make any changes to the kinesis exporter itself
and tried to keep it mostly a 1-1 translation from OC to OT. I think we
can move the kinesis exporter code into this repo once we migrate
completely to OT internally.