Remove workers from octrace receiver (#13)
* Remove workers from octrace receiver

These workers were added sometime ago in an attempt to increase ingestion rate, however, they cause the loss of the error returned by the next consumer on the pipeline. Measurements of typical scenarios do not indicate reduction in ingress rate but if some async like the removed workers is detect it should be added via a pipeline processor. This allows a pipeline processor like memory-limiter to affect the response of a receiver while having workers/async behavior down in the pipeline.

* Add description for new test